Refactored code
Now, instead of one library and one
binary, it's two libraries and one binary
