+++ Cargo.toml
@@ -1,16 +1,6 @@
-[package]
-name = "tego"
-version = "0.4.2"
-authors = ["Brendon Bown <brendon.bown@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-nom = "5.0.0"
-structopt = "0.3.14"
-owned_chars = "0.3.1"
-
-[dev-dependencies]
-tempfile = "3.1.0"
-assert_cmd = "1"
+[workspace]
+members = [
+	"tego_parser",
+	"tego_interpreter",
+	"tego"
+]
