--- src/ast/mod.rs
@@ -1,10 +1,0 @@
-mod decl;
-mod expr;
-mod match_;
-mod prog;
-
-pub use decl::Decl;
-pub use expr::Expr;
-pub use match_::Match;
-pub use match_::MatchVal;
-pub use prog::Prog;
