+++ tego_interpreter/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "tego_interpreter"
+version = "0.1.0"
+authors = ["Brendon Bown <bhbochikens@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+owned_chars = "0.3.1"
+tego_parser = { path = "../tego_parser" }
