+++ tego_interpreter/src/lib.rs
@@ -1,0 +1,18 @@
+#[allow(unused_macros)]
+macro_rules! basic_test {
+    ( $name:ident $( $actual:expr => $expected:expr );+) => {
+        #[allow(clippy::eq_op)]
+        #[test]
+        fn $name() {
+            $( assert_eq!($expected, $actual); )+
+        }
+    };
+}
+
+mod interpreter;
+mod environment;
+mod type_;
+pub mod prelude;
+pub mod value;
+
+pub use interpreter::*;
