+++ tego_parser/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tego_parser"
+version = "0.1.0"
+authors = ["Brendon Bown <bhbochikens@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+nom = "5.0.0"
