+++ tego_parser/src/parsers.rs
@@ -1,0 +1,5 @@
+pub mod decl;
+pub mod expr;
+pub mod match_;
+pub mod prog;
+pub mod tokens;
