+++ tego_parser/src/span.rs
